requirements.txt
Flask
pytesseract
PyMuPDF
Pillow

